GCs
GCs (Globular Clusters) is a project to estimate the mass and cumulative mass profiles of globular clusters. Within a Bayesian framework, the position and velocity data of stars in a GC are used to obtain a posterior distribution for the relevant model parameters in the [LIMEPY and SPES models](https://github.com/mgieles/limepy).