speagle python scripts
Additional Python scripts for working with GC modelling with `limepy` in Python. A summary of these changes is below.

Data generation:
`model_gen_v2.py` and `model_regen_v2.py` are scripts that are patterned off of Gwen's original scripts, but with some additional flexibility to allow for anisotropic models.

Optimization:
`model_optimize.py` loads in the data, subsamples it (by default to `logn=2.7` or `n=501` stars), and then optimizes the result. There are options to allow for anisotropic models and more/less robust optimization options (differential evolution vs Nelder-Mead). This flexibility allows the user to explore how robust the optimization is, and also cases where the data are correctly modelled (data=iso, model=iso; data=aniso, model=aniso), incorrectly modelled (data=aniso, model=iso), or involve an overly-complex model (data=iso, model=aniso). This saves the MAP solution (position and log-probability) as well as the Hessian (with some safety checks to ensure it is semi-positive definite).

Importance Sampling:
`model_IS.py` reads in the results from `model_optimize.py` and performs a very simple Gaussian proposal (where the size of the covariance is just inflated relative to the default Hessian). The mean and covariance of the proposal along with samples and log-probabilities (of the proposal and posterior) are then saved. A diagnostic set of corner plots (using `corner`) are also generated.

MCMC:
`model_MCMC.py` uses the results from `model_IS.py` to sample using MCMC. This builds off of `emcee` using the Differential Evolution ensemble moves. The user can adjust the number of chains, burn-in iterations, and sampling iterations. Samples and log-probabilities (both initial Gaussian proposal and posterior) are then saved along with a diagnostic corner plot.

Nested Sampling:
`model_NS.py` uses Nested Sampling to estimate the posterior and the marginal likelihood (evidence). No additional inputs from previous modelling steps are required. This runs off of `dynesty`. The user can change the number of live points and the point at which the sampler begins to construct bounding distributions. Samples, log-weights, and the log-evidence (plus errors) are then saved along with a diagnostic plot.

While the code has been tested on a simple example, it has not yet been stress-tested on a variety of parameters or model cases.