inflate bugfix
Fixes a bug where the Hessian ignored the `inflate` factor
when trying to limit overly large axes to remain within the
confines of the prior distribution. This didn't matter
prevously when such points were clipped (and the distribution's
normalization constant incorrectly estimated numerically)
but became an issue once *that* particular bug was removed.

Also adjusted the default `inflate` factor down from 2.5 to 2.